renamed login form component (fix)
