+++ debug.log
@@ -18,3 +18,4 @@
 [0405/001232.736:ERROR:crash_report_database_win.cc(428)] unexpected header
 [0405/024854.264:ERROR:crash_report_database_win.cc(428)] unexpected header
 [0405/031439.663:ERROR:crash_report_database_win.cc(428)] unexpected header
+[0405/035041.323:ERROR:crash_report_database_win.cc(428)] unexpected header
